Create KubernetesClusterScript
This script will first execute a series of commands on the master node to install and configure Kubernetes. Then, it will get the join command for the worker nodes from the master node and execute that command on each worker node to join it to the Kubernetes cluster.
Please keep in mind that this is just a sample script and you may need to modify it to suit your specific requirements and environment.